+++ client-database-android/src/test/java/de/njsm/stocks/client/database/TypeConvertersBigDecimalTest.java
@@ -61,4 +61,10 @@ public class TypeConvertersBigDecimalTest {
     public void converterPreservesIdentity() {
         assertEquals(input, uut.dbToBigDecimal(uut.bigDecimalToDb(input)));
     }
+
+    @Test
+    public void exampleWorks() {
+        String input = "1.23";
+        assertEquals(new BigDecimal(input), uut.dbToBigDecimal(input));
+    }
 }